refactored buttonwithwarning
